+++ MaxMind.MinFraud/Request/Transaction.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace MaxMind.MinFraud.Request
@@ -53,26 +52,6 @@ namespace MaxMind.MinFraud.Request
             Payment = payment;
             Shipping = shipping;
             ShoppingCart = shoppingCart;
-        }
-
-        /// <summary>
-        ///     This constructor only exists for backward compatability.
-        /// </summary>
-        [Obsolete]
-        public Transaction(
-            Device device,
-            Account account,
-            Billing billing,
-            CreditCard creditCard,
-            Email email,
-            Event userEvent,
-            Order order,
-            Payment payment,
-            Shipping shipping,
-            IList<ShoppingCartItem> shoppingCart
-        ) : this(device, account, billing, creditCard, null, email,
-            userEvent, order, payment, shipping, shoppingCart)
-        {
         }
 
         /// <summary>